Rename ControlUnitAdministration to Administration in Backend
