+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/entities/administration/Administration.kt
@@ -1,0 +1,7 @@
+package fr.gouv.cnsp.monitorfish.domain.entities.administration
+
+data class Administration(
+    val id: Int,
+    val isArchived: Boolean,
+    val name: String,
+)
