--- backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/entities/control_unit/ControlUnitAdministration.kt
@@ -1,7 +1,0 @@
-package fr.gouv.cnsp.monitorfish.domain.entities.control_unit
-
-data class ControlUnitAdministration(
-    val id: Int,
-    val isArchived: Boolean,
-    val name: String,
-)
