+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/entities/control_unit/ControlUnitDepartmentArea.kt
@@ -1,5 +1,6 @@
 package fr.gouv.cnsp.monitorfish.domain.entities.control_unit
 
+// TODO Maybe merge `districts` from MonitorFish into MonitorEnv `departmentAreas`?
 data class ControlUnitDepartmentArea(
     /** `inseeCode` is the ID. */
     val inseeCode: String,
