+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/use_cases/control_units/dtos/FullControlUnit.kt
@@ -1,14 +1,13 @@
 package fr.gouv.cnsp.monitorfish.domain.use_cases.control_units.dtos
 
-import fr.gouv.cnsp.monitorfish.domain.entities.control_unit.ControlUnitAdministration
+import fr.gouv.cnsp.monitorfish.domain.entities.administration.Administration
 import fr.gouv.cnsp.monitorfish.domain.entities.control_unit.ControlUnitContact
 import fr.gouv.cnsp.monitorfish.domain.entities.control_unit.ControlUnitDepartmentArea
-import fr.gouv.cnsp.monitorfish.domain.entities.control_unit.FullControlUnitResource
 
 data class FullControlUnit(
     val id: Int,
     val areaNote: String?,
-    val administration: ControlUnitAdministration,
+    val administration: Administration,
     val administrationId: Int,
     val controlUnitContactIds: List<Int>,
     val controlUnitContacts: List<ControlUnitContact>,
