+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/use_cases/control_units/dtos/FullControlUnitResource.kt
@@ -1,4 +1,8 @@
-package fr.gouv.cnsp.monitorfish.domain.entities.control_unit
+package fr.gouv.cnsp.monitorfish.domain.use_cases.control_units.dtos
+
+import fr.gouv.cnsp.monitorfish.domain.entities.control_unit.ControlUnit
+import fr.gouv.cnsp.monitorfish.domain.entities.control_unit.ControlUnitResourceType
+import fr.gouv.cnsp.monitorfish.domain.entities.control_unit.ControlUnitStation
 
 data class FullControlUnitResource(
     val id: Int,
